@@ -22,7 +22,7 @@
         }
     },
     "require": {
-        "php": ">=7.3",
+        "php": "^8.1",
         "ext-dom": "*",
         "ext-openssl": "*",
         "eclipxe/enum": "^0.2",
@@ -30,7 +30,7 @@
     },
     "require-dev": {
         "robrichards/xmlseclibs": "^3.1.0",
-        "phpunit/phpunit": "^9.5"
+        "phpunit/phpunit": "^10.5.46"
     },
     "suggest": {
         "robrichards/xmlseclibs": "Create document signatures (partially) using xmlseclibs"
@@ -57,12 +57,12 @@
         ],
         "dev:test": [
             "@dev:check-style",
-            "@php vendor/bin/phpunit --testdox --verbose --stop-on-failure",
+            "@php vendor/bin/phpunit --testdox --display-all-issues --stop-on-failure",
             "@php tools/phpstan analyse --no-progress",
             "@php tools/infection --no-progress --no-interaction --show-mutations"
         ],
         "dev:coverage": [
-            "@php -dzend_extension=xdebug.so -dxdebug.mode=coverage vendor/bin/phpunit --verbose --coverage-html build/coverage/html/"
+            "@php -dzend_extension=xdebug.so -dxdebug.mode=coverage vendor/bin/phpunit --coverage-html build/coverage/html/"
         ]
     },
     "scripts-descriptions": {
