@@ -15,19 +15,20 @@ use PhpCfdi\XmlCancelacion\Signers\DOMSigner;
 use PhpCfdi\XmlCancelacion\Signers\SignerInterface;
 use PhpCfdi\XmlCancelacion\Tests\TestCase;
 use PhpCfdi\XmlCancelacion\XmlCancelacionHelper;
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\MockObject\MockObject;
 
-/** @covers \PhpCfdi\XmlCancelacion\XmlCancelacionHelper */
+#[CoversClass(XmlCancelacionHelper::class)]
 final class XmlCancelacionHelperTest extends TestCase
 {
     /** @return Credentials&MockObject */
-    private function createFakeCredentials()
+    private function createFakeCredentials(): Credentials
     {
         return $this->createMock(Credentials::class);
     }
 
     /** @return  SignerInterface&MockObject $fakeSigner */
-    private function createFakeSigner()
+    private function createFakeSigner(): SignerInterface
     {
         return $this->createMock(SignerInterface::class);
     }
